requirements.txt
Flask==3.0.2
python-dotenv==1.0.0
pandas==2.1.4
